Settings: use different dependency to not need QUERY_ALL_PACKAGES
Google disallows QUERY_ALL_PACKAGES permission. The old library
installed_apps  was requesting this permission automatically,
unfortunately. Now switched to device_apps which doesn't do this, and
included a list of well-known banking apps to the <queries> section to
read their information. Unfortunately, device_apps is unable to submit
the app icon. In the future, a fork of installed_apps might be better
suited.