Show AppLogo during biometrics check
If user resumes the app after some time, a biometrics check is required.
Until now, the last screen was shown during this check, which might
contain sensitive data that could be extracted via screenshots.
This commit introduces a new screen with just the app logo that is shown
during the biometrics check.

fixes #585